reorganizing calibration to incorporate Lu's work from bdai_calibration
